api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
psycopg[binary]==3.2.9
pgvector==0.4.1
openai==1.35.7
python-dotenv==1.0.1
httpx==0.27.2
bcrypt==4.2.0
PyJWT==2.9.0